shseC didn't have correct type for nlon. Minor changes to error messages.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@12068 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -2824,7 +2824,7 @@ void NclAddUserFuncs(void)
     nargs = 0;
     args = NewArgs(2);
     SetArgTemplate(args,nargs,"numeric",0,NclANY);nargs++;
-    SetArgTemplate(args,nargs,"integer",1,NclANY);nargs++;
+    SetArgTemplate(args,nargs,"numeric",1,NclANY);nargs++;
 
     NclRegisterFunc(shseC_W,args,"shseC",nargs);
